pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calc_notifier_telegram"
version = "0.1.0"
description = "Минималистичный Telegram-only notifier для расчётов: PDF, картинки, файлы, история (3 последних отчёта в чате)."
authors = [{name="Artyom Loginov", email="loginov.artem@phystech.edu"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "matplotlib>=3.8",
  "Pillow>=9.5",
  "reportlab>=4.0",
  "colorama>=0.4.6"
]
